refactor: extract AgentPicker + useForgeAgents into its own file
Step 1 of specs/SPEC_AGENT_VIEW_MODULARIZATION_2026_04_13.md. The
goal of the full refactor is to bring agent-view.tsx from 1,178
lines down to ≤300 by extracting single-concern hooks + components.
This commit does the first extraction.

Scope:
  - New file: frontend/app/view/agent/components/AgentPicker.tsx
    * useForgeAgents hook (subscribes to forgeagents:changed, reactive
      accessor for the current Forge list)
    * AgentPicker component (the empty-state + card-list UI shown when
      the block has no agentId in meta; launches the selected agent via
      model.launchForgeAgent)
  - agent-view.tsx:
    * Remove the inline useForgeAgents function
    * Remove the inline AgentPicker component
    * Add `import { AgentPicker } from "./components/AgentPicker"`
    * No other changes — the AgentViewWrapper Show/Fallback still
      renders <AgentPicker model={model} /> as before

Net:
  - agent-view.tsx: 1,178 → 1,053 lines (−125)
  - AgentPicker.tsx: new, 151 lines (+18 net vs what was extracted,
    all from import boilerplate and the header comment block)

This is the smallest useful refactor step — zero behavior change,
zero risk. AgentPicker was already well-isolated inside agent-view.tsx
(it didn't share any state with AgentPresentationView), so the move
is mechanical.

SolidJS reactivity note: I converted AgentPicker to use
`(props: Props)` accessor pattern instead of the destructured
`({ model }: { model: ... })` it had inline. Destructuring component
props is the exact trap that bit ToolBlock in PR #346 — even though
`model` happens to be a stable reference here (it's a class
instance, not a Solid signal), using props.model everywhere keeps
the code consistent with the house rule.

Next steps per the spec:
  2. Launch flow extraction (flows/launch-flow.ts)
  3. useLaunchLogs hook
  4. useAgentControllerStatus hook
  5. useHistoryPagination hook
  6. useSessionDigest hook
  7. useBookmarks hook
  8. useInSessionSearch hook
  9. useScrollToNode hook  ← also kills the scrollIntoView bug (already
     fixed standalone in PR #347)
  10. useAgentKeyboard hook
  11. AgentPresentationHeader component
  12. Cleanup pass, target agent-view.tsx ≤ 300 lines
